+++ src/app/page.tsx
@@ -1,7 +1,5 @@
-import Footer from '@/components/Footer';
-import Header from '@/components/Header';
-import Layoutheader from '@/components/Layoutheader';
-import Main from '@/components/Main';
+import Header from '@/components/MainPage/Header';
+import Main from '@/components/MainPage/Main';
 
 export default function Home() {
   return (
