--- src/components/Localsection/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Localsection';
