+++ src/components/MainPage/Header/Header.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import React, { useState } from 'react';
+import React from 'react';
 
 export default function Header({ title, content }) {
   return (
