+++ src/components/MainPage/Main/index.ts
