+++ src/components/MainPage/Main/section/LocalSection.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
 
-export default function Localsection() {
+export default function LocalSection() {
   return <section></section>;
 }
