--- src/components/SearchItem/index.ts
@@ -1,1 +1,0 @@
-export { default } from './SearchItem';
